fix multi connector build

+++ setup.py
@@ -11,7 +11,6 @@ from jsonmerge import merge
 import tempfile
 
 here = os.path.abspath(os.path.dirname(__file__))
-cleanup_file_list = []
 
 with open(os.path.join(here, 'README.md'), encoding='utf-8') as f:
     long_description = f.read()
@@ -66,6 +65,7 @@ else:
         exit(1)
 
 for project_name in projects.keys():
+    cleanup_file_list = []
     temp_dir = None
     module_dir = None
 
@@ -144,34 +144,34 @@ for project_name in projects.keys():
     shutil.rmtree('MANIFEST.in', ignore_errors=True)
     shutil.copyfile('build_templates/MANIFEST.in', 'MANIFEST.in')
     json_include_lines = []
+
     for json_search_path in src_folders:
         module_dir = json_search_path
-        conf_path = os.path.join(module_dir, 'conf')
-        shutil.rmtree(conf_path, ignore_errors=True)
-        os.mkdir(conf_path)
         configuration_path = os.path.join(module_dir, 'configuration')
-        for r, d, f in os.walk(configuration_path):
-            for file in f:
-                configuration_file_path = os.path.join(r, file)
-                with open(configuration_file_path) as json_file:
-                    module_data = json.load(json_file)
-                base_data = None
-                data = dict()
-                base_path = os.path.join(r, '..', '..', file)
-                if os.path.isfile(base_path):
-                    with open(base_path) as json_file:
-                        base_data = json.load(json_file)
-                    data = base_data
-                data = merge(data, module_data)
-
-                json_file_path = os.path.join(r, '..', 'conf', file)
-                with open(json_file_path, 'w') as json_file:
-                    json_file.write(json.dumps(data, indent=4, sort_keys=False))
-
-        temp_dir = tempfile.TemporaryDirectory()
-        shutil.move(configuration_path, temp_dir.name)
-        os.rename(conf_path, configuration_path)
-        cleanup_file_list.append(configuration_path)
+        if os.path.isdir(configuration_path):
+            conf_path = os.path.join(module_dir, 'conf')
+            shutil.rmtree(conf_path, ignore_errors=True)
+            os.mkdir(conf_path)
+            for r, d, f in os.walk(configuration_path):
+                for file in f:
+                    configuration_file_path = os.path.join(r, file)
+                    with open(configuration_file_path) as json_file:
+                        module_data = json.load(json_file)
+                    base_data = None
+                    data = dict()
+                    base_path = os.path.join(r, '..', '..', file)
+                    if os.path.isfile(base_path):
+                        with open(base_path) as json_file:
+                            base_data = json.load(json_file)
+                        data = base_data
+                    data = merge(data, module_data)
+                    json_file_path = os.path.join(r, '..', 'conf', file)
+                    with open(json_file_path, 'w') as json_file:
+                        json_file.write(json.dumps(data, indent=4, sort_keys=False))
+            temp_dir = tempfile.TemporaryDirectory()
+            shutil.move(configuration_path, temp_dir.name)
+            os.rename(conf_path, configuration_path)
+            cleanup_file_list.append(configuration_path)
 
         for r, d, f in os.walk(module_dir):
             r_split = r.split(os.sep)
@@ -200,5 +200,7 @@ for project_name in projects.keys():
                 shutil.rmtree(cleanup_file)
             else:
                 os.remove(cleanup_file)
-    shutil.move(os.path.join(temp_dir.name, 'configuration'), module_dir)
+    if temp_dir is not None:
+        shutil.move(os.path.join(temp_dir.name, 'configuration'), module_dir)
+        temp_dir = None
     print('---------------------------------')
